+++ src/main/java/org/kumoricon/registration/settings/SettingsService.java
@@ -1,0 +1,67 @@
+package org.kumoricon.registration.settings;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import javax.annotation.PostConstruct;
+
+/**
+ * This service provides configuration options that can be set at run time. Default values may
+ * be loaded from the configuration file (usually `application.properties`), but if the settings
+ * are also in the database the values in the database will be used.
+ *
+ * Settings are cached in memory (in an immutable object) to save a database round trip
+ */
+@Service
+public class SettingsService {
+    private static final Logger log = LoggerFactory.getLogger(SettingsService.class);
+    private final SettingsRepository settingsRepository;
+    private Settings currentSettings;
+
+    // Default startup values from config file; will be overwritten by values stored in the database
+    @Value("${kumoreg.printing.enablePrintingFromServer}")
+    private Boolean enablePrintingFromServerDefault;
+    @Value("${kumoreg.trainingMode}")
+    private Boolean trainingModeDefault;
+
+    public SettingsService(SettingsRepository settingsRepository) {
+        this.settingsRepository = settingsRepository;
+    }
+
+    public Settings getCurrentSettings() {
+        return currentSettings;
+    }
+
+    @PostConstruct
+    public void init() {
+        log.info("Loading settings from database...");
+        Settings.Builder settings = defaultSettingsBuilder();
+        settingsRepository.loadSettingsInto(settings);
+        currentSettings = settings.build();
+    }
+
+    private Settings.Builder defaultSettingsBuilder() {
+        return new Settings.Builder()
+                .setTrainingMode(trainingModeDefault)
+                .setEnablePrinting(enablePrintingFromServerDefault)
+                .setReportPrinterName(null)
+                .setUpdated(System.currentTimeMillis());
+    }
+
+    public void setPrintingEnabled(Boolean value) {
+        settingsRepository.upsertSetting(SettingsRepository.ENABLE_PRINTING, value.toString());
+        this.currentSettings = new Settings.Builder(currentSettings).setEnablePrinting(value).build();
+    }
+
+    public void setTrainingMode(Boolean value) {
+        settingsRepository.upsertSetting(SettingsRepository.TRAINING_MODE, value.toString());
+        this.currentSettings = new Settings.Builder(currentSettings).setTrainingMode(value).build();
+    }
+
+    public void setReportPrinterName(String printerName) {
+        this.currentSettings = new Settings.Builder(currentSettings).setReportPrinterName(printerName).build();
+        settingsRepository.upsertSetting(SettingsRepository.REPORT_PRINTER_NAME, printerName);
+    }
+}