+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Change History
 Version 1.0.2 (pending)
 - Search staff list server-side for performance on slow clients
+- Added each user's total logged in time for the displayed day to the Login History report
+- Time period is displayed when you hover over a cell in the Login History Report
 
 Version 1.0.1 (11/14/2019)
 - Staff check in: don't search staff list until 2 characters have been entered in search (for performance)