+++ CHANGES.md
@@ -1,5 +1,5 @@
 # Change History
-Version 1.0.2 (pending)
+Version 1.1.0 (pending)
 - Search staff list server-side for performance on slow clients
 - Added each user's total logged in time for the displayed day to the Login History report
 - Time period is displayed when you hover over a cell in the Login History Report
@@ -7,6 +7,7 @@ Version 1.0.2 (pending)
 - Switch to Jetty from Tomcat, which increased performance, especially for file uploads
 - Upload staff photos as JPG instead of PNG
 - Save staff photos and signatures with the name {name}-{id}-photo/signature-{timestamp}.jpg/png
+- Added configurable run-time settings (schema change) and configuration screen
 
 Version 1.0.1 (11/14/2019)
 - Staff check in: don't search staff list until 2 characters have been entered in search (for performance)